StorageBuckets: Bucket Directory Deletion
This change updates QuotaDatabase::DeleteBucketData
(previously QuotaDatabase::DeleteBucketInfo) to delete
the new bucket directory path [1] on bucket deletion
for in-memory.

Some design decisions worth mentioning:

* Storage path construction is now extracted into a new
  utils file used by StorageDirectory and QuotaManagerProxy.
  Ideally this would be only used by StorageDirectory,
  but would require lots restructuring around how Storage APIs
  are implemented now. Having the path method available
  in QuotaManagerProxy is beneficial since it would allow
  Storage APIs to build path without thread hopping.
  This is especially important for Storage APIs that use this
  because retrieving this is part of the critical path for
  storage usage.

* Doom before delete. Throughout StorageDirectory we choose
  to invalidate a path by renaming before deletion. This is
  because deletion performs differently per OS and tends to
  have a higher failure rate particularly on Windows.
  We want to make sure the data is inaccessible immediately
  and clean-up after all other removal has succeeded.

[1][Googlers] https://docs.google.com/document/d/1D9Euw9ZgMw2YADFcuOmGj1QAT3e-4V_K6y8MMHxofEQ/edit?resourcekey=0-MQc2XFLLf7Nl3cWdTY3qtg#heading=h.mloycxf0nud5

Bug: 1315386
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3586243
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#995840}

+++ storage/browser/quota/quota_database.cc
@@ -558,16 +558,21 @@ QuotaError QuotaDatabase::DeleteHostQuota(const std::string& host,
   return QuotaError::kNone;
 }
 
-QuotaError QuotaDatabase::DeleteBucketInfo(BucketId bucket_id) {
+QuotaError QuotaDatabase::DeleteBucketData(const BucketLocator& bucket) {
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
-  DCHECK(!bucket_id.is_null());
   QuotaError open_error = EnsureOpened();
   if (open_error != QuotaError::kNone)
     return open_error;
 
+  // Doom bucket directory first so data is no longer accessible, even if
+  // directory deletion fails. `storage_directory_` may be nullptr for
+  // in-memory only.
+  if (storage_directory_ && !storage_directory_->DoomBucket(bucket))
+    return QuotaError::kFileOperationError;
+
   static constexpr char kSql[] = "DELETE FROM buckets WHERE id = ?";
   sql::Statement statement(db_->GetCachedStatement(SQL_FROM_HERE, kSql));
-  statement.BindInt64(0, bucket_id.value());
+  statement.BindInt64(0, bucket.id.value());
 
   if (!statement.Run())
     return QuotaError::kDatabaseError;
@@ -584,6 +589,9 @@ QuotaError QuotaDatabase::DeleteBucketInfo(BucketId bucket_id) {
   // TODO(crbug.com/1314567): For handling inconsistencies between the db and
   // the file system.
   ScheduleCommit();
+
+  if (storage_directory_)
+    storage_directory_->ClearDoomedBuckets();
 
   return QuotaError::kNone;
 }
